Added some Utilities

+++ src/main/java/de/tobiyas/util/inventorymenu/stats/IntegerSelectionInterface.java
@@ -156,6 +156,6 @@ public class IntegerSelectionInterface extends AbstractStatSelectionInterface {
 
 	@Override
 	protected Object unparseValue() {
-		return value;
+		return (int) value;
 	}
 }
